requirements.txt
flask==3.1.0
google-genai==1.3.0
gunicorn==23.0.0
